refactor(auto-reply): keep re-promotion selector off the public plugin-SDK surface
Move AutoFallbackModelRef and resolveAutoFallbackRepromotionTarget out of
src/agents/agent-scope.ts (transitively public via plugin-sdk) into
src/auto-reply/reply/auto-fallback-repromotion.ts, which is not on the
public surface. Pure relocation; no behavior change. Restores the
plugin-sdk export budget pin without bumping the baseline.
